requirements.txt
AfricastalkingGateway==1.7
alembic==0.9.1
click==6.7
dominate==2.3.1
Flask==0.12
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==1.0
migrate==0.3.8
python-editor==1.0.3
requests==2.13.0
SQLAlchemy==1.1.6
visitor==0.1.3
Werkzeug==0.12.1
gunicorn